Remove Signal Search Hierarcy dialog
