+++ man/gtkwaverc.5
@@ -244,9 +244,6 @@ Hides the Signal Search Tree widget for GTK2.4 and greater such that it is not e
 \fBhier_delimeter\fR <\fIvalue\fP>
 This allows characters other than '/' to be used to delimit levels in the hierarchy. Only the first character in the value is significant.
 .TP 
-\fBhier_grouping\fR <\fIvalue\fP>
-For the tree widgets, this allows the hierarchies to be grouped in a single place rather than spread among the netnames.
-.TP 
 \fBhier_ignore_escapes\fR <\fIvalue\fP>
 A nonzero value indicates that the signal pane ignores escapes in identifiers when determining the hierarchy maximum depth.  Default is disabled so that escapes are examined.
 .TP
