+++ src/gtk23compat.h
@@ -26,7 +26,6 @@
 /* seat vs pointer enable */
 #define WAVE_ALLOW_GTK3_SEAT_VS_POINTER_GRAB_UNGRAB
 
-#define WAVE_GTK3_HIERSEARCH_DEBOUNCE
 #define WAVE_GTK3_MENU_SEPARATOR
 #define WAVE_ALLOW_GTK3_CAIRO_CREATE_FIX
