+++ src/menu.h
@@ -201,7 +201,6 @@ enum WV_MenuItems
     WV_MENU_SPS2,
     WV_MENU_SEP7B,
     WV_MENU_SSR,
-    WV_MENU_SSH,
     WV_MENU_SST,
     WV_MENU_SEP7,
 #if !defined __MINGW32__
@@ -213,7 +212,6 @@ enum WV_MenuItems
     WV_MENU_ACOL,
     WV_MENU_ACOLR,
     WV_MENU_ABON,
-    WV_MENU_HTGP,
     WV_MENU_SEP7C,
     WV_MENU_STRSE,
     WV_MENU_TMTT,
@@ -375,12 +373,10 @@ void menu_lexize(gpointer null_data, guint callback_action, GtkWidget *widget);
 void menu_reverse(gpointer null_data, guint callback_action, GtkWidget *widget);
 void menu_tracesearchbox(gpointer null_data, guint callback_action, GtkWidget *widget);
 void menu_signalsearch(gpointer null_data, guint callback_action, GtkWidget *widget);
-void menu_hiersearch(gpointer null_data, guint callback_action, GtkWidget *widget);
 void menu_treesearch(gpointer null_data, guint callback_action, GtkWidget *widget);
 void menu_autocoalesce(gpointer null_data, guint callback_action, GtkWidget *widget);
 void menu_autocoalesce_reversal(gpointer null_data, guint callback_action, GtkWidget *widget);
 void menu_autoname_bundles_on(gpointer null_data, guint callback_action, GtkWidget *widget);
-void menu_hgrouping(gpointer null_data, guint callback_action, GtkWidget *widget);
 void menu_movetotime(gpointer null_data, guint callback_action, GtkWidget *widget);
 void menu_zoomsize(gpointer null_data, guint callback_action, GtkWidget *widget);
 void menu_zoombase(gpointer null_data, guint callback_action, GtkWidget *widget);
