+++ src/rc.h
@@ -92,7 +92,6 @@ int  f_fontname_waves (char *str);
 int  f_force_toolbars (char *str);
 int  f_hide_sst (char *str);
 int  f_hier_delimeter (char *str);
-int  f_hier_grouping (char *str);
 int  f_hier_max_level (char *str);
 int  f_hpane_pack (char *str);
 int  f_ignore_savefile_pos (char *str);
