+++ src/tree.c
@@ -1067,3 +1067,52 @@ if(GLOBALS->sst_exclude_filename)
 	}
 
 }
+
+/* Get the highest signal from T.  */
+struct tree *fetchhigh(struct tree *t)
+{
+    while (t->child)
+        t = t->child;
+    return (t);
+}
+
+/* Get the lowest signal from T.  */
+struct tree *fetchlow(struct tree *t)
+{
+    if (t->child) {
+        t = t->child;
+
+        for (;;) {
+            while (t->next)
+                t = t->next;
+            if (t->child)
+                t = t->child;
+            else
+                break;
+        }
+    }
+    return (t);
+}
+
+void recurse_fetch_high_low(struct tree *t)
+{
+top:
+    if (t->t_which >= 0) {
+        if (t->t_which > GLOBALS->fetchhigh)
+            GLOBALS->fetchhigh = t->t_which;
+        if (GLOBALS->fetchlow < 0) {
+            GLOBALS->fetchlow = t->t_which;
+        } else if (t->t_which < GLOBALS->fetchlow) {
+            GLOBALS->fetchlow = t->t_which;
+        }
+    }
+
+    if (t->child) {
+        recurse_fetch_high_low(t->child);
+    }
+
+    if (t->next) {
+        t = t->next;
+        goto top;
+    }
+}
